Reduce splint warnings

+++ bytes.c
@@ -2,6 +2,6 @@
 
 unsigned short be16toh_array(unsigned char* array)
 {
-    unsigned short ret = (array[0] << 8) | (array[1]);
+    unsigned short ret = (unsigned short)((array[0] << 8) | (array[1]));
     return ret;
 }
